+++ core-bundle/src/Resources/contao/elements/ContentDownloads.php
@@ -175,7 +175,7 @@ class ContentDownloads extends \ContentElement
 			// Folders
 			else
 			{
-				$objSubfiles = \FilesModel::findByPid($objFiles->uuid);
+				$objSubfiles = \FilesModel::findByPid($objFiles->uuid, array('order' => 'name'));
 
 				if ($objSubfiles === null)
 				{